Guest controller  ->remove constructor
 Web -> add welcome page to guest middleware

 Create session alerts folder with session messages, remove duplicate code use

 Rename "Back to admin page" buttons to "Back to Home"

 User edit page now use auth()->user data

 Invoice quantity from integer to double

 Home page -> replace welcome message

+++ database/migrations/2019_10_31_121328_create_invoices_table.php
@@ -17,7 +17,7 @@ class CreateInvoicesTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedInteger('number');
             $table->string('product_name');
-            $table->unsignedInteger('quantity');
+            $table->double('quantity');
             $table->set('unit', ['kg', 'grams', 'qty.', 'cm', 'liters']);
             $table->decimal('unit_price');
             $table->unsignedBigInteger('added_by');
